requirements.txt
amqp==2.1.4
amqplib==1.0.2
aniso8601==1.1.0
anyjson==0.3.3
Babel==2.2.0
backports.ssl-match-hostname==3.4.0.2
bcrypt==3.1.4
billiard==3.3.0.23
boto==2.48.0
certifi==2016.2.28
cffi==1.11.5
chardet==2.2.1
configobj==4.7.2
configparser==3.5.0
decorator==3.4.0
dnspython==1.13.0
ecdsa==0.13
enum34==1.1.2
Flask==0.10.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
future==0.15.2
futures==3.0.5
google-compute-engine==2.3.5
idna==2.0
inflect==0.2.5
iniparse==0.4
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
#perf==0.1
phpserialize==1.3
pika==0.10.0
pyasn1==0.1.9
pycparser==2.14
PyMySQL==0.7.1
python-dateutil==2.4.2
python-jose==3.0.0
pytz==2015.7
pyudev==0.15
redis==2.10.6
requests==2.9.1
rsa==3.4.1
six==1.10.0
sqlacodegen==1.1.6
SQLAlchemy==1.0.11
urllib3==1.10.2
vine==1.1.3
virtualenv==15.1.0
Werkzeug==0.11.5
xlrd==0.9.4
xmltodict==0.10.1
